requirements.txt
discord.py>=2.3.2
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
